design update

+++ Models/ShoppingCart.cs
@@ -105,7 +105,7 @@ namespace NetStore.Models
             _appDbContext.SaveChanges();
         }
 
-        public decimal GetShoppingCartTotal()
+        public double GetShoppingCartTotal()
         {
             var total = _appDbContext.ShoppingCartItems.Where(c => c.ShoppingCartId == ShoppingCartId)
                 .Select(c => c.Book.Price * c.Amount).Sum();
